refactor: move _check_po_rfq_files to BaseModule; cover quote module
- _check_po_rfq_files promoted to BaseModule so Job and Quote share one
  implementation with no duplication
- Quote module now hooks all three file-add paths: Create New drop zone,
  Link Drawings dialog, and Add to Existing drop zone
- Job module's duplicate method and its now-unused QMessageBox /
  classify_document imports removed
