+++ core/base_module.py
@@ -6,7 +6,9 @@ This enables dynamic loading and ensures consistent behavior across modules.
 """
 
 from abc import ABC, abstractmethod
-from PyQt6.QtWidgets import QWidget
+from pathlib import Path
+from typing import List
+from PyQt6.QtWidgets import QMessageBox, QWidget
 
 
 class BaseModule(ABC):
@@ -158,3 +160,35 @@ class BaseModule(ABC):
         """
         if self._app_context:
             self._app_context.show_info(title, message)
+
+    def _check_po_rfq_files(self, newly_added: List[str], files_store: List[str], list_widget) -> None:
+        """Scan newly-added files for PO/RFQ signals and prompt the user for each flagged file.
+
+        Only files whose extension matches the configured blueprint extensions are checked —
+        non-drawing files (.msg, .docx, etc.) are always documents and need no prompt.
+        """
+        from shared.utils import classify_document
+        bp_exts = [e.lower() for e in self.app_context.get_setting('blueprint_extensions', ['.pdf', '.dwg', '.dxf'])]
+        for path in list(newly_added):
+            if Path(path).suffix.lower() not in bp_exts:
+                continue
+            is_flagged, reason = classify_document(path)
+            if not is_flagged:
+                continue
+            filename = Path(path).name
+            msg = QMessageBox(self._widget)
+            msg.setWindowTitle("PO / RFQ Detected")
+            msg.setIcon(QMessageBox.Icon.Warning)
+            msg.setText(f"'{filename}' appears to be a Purchase Order or RFQ.")
+            msg.setInformativeText(
+                f"Detected: {reason}\n\n"
+                "Save it as a document in the job folder, or remove it from the list?"
+            )
+            save_btn = msg.addButton("Save as Document", QMessageBox.ButtonRole.AcceptRole)
+            remove_btn = msg.addButton("Remove", QMessageBox.ButtonRole.RejectRole)
+            msg.setDefaultButton(save_btn)
+            msg.exec()
+            if msg.clickedButton() is remove_btn and path in files_store:
+                idx = files_store.index(path)
+                files_store.pop(idx)
+                list_widget.takeItem(idx)
