+++ modules/job/module.py
@@ -13,7 +13,7 @@ import shutil
 from pathlib import Path
 from typing import List
 from PyQt6.QtWidgets import (
-    QWidget, QTreeWidgetItem, QButtonGroup, QAbstractItemView, QMessageBox
+    QWidget, QTreeWidgetItem, QButtonGroup, QAbstractItemView
 )
 
 from PyQt6.QtCore import Qt, QThread, pyqtSignal
@@ -27,7 +27,7 @@ from shared.widgets import (
 )
 from shared.utils import (
     is_blueprint_file, parse_job_numbers, create_file_link,
-    sanitize_filename, open_folder, get_next_number, classify_document
+    sanitize_filename, open_folder, get_next_number
 )
 
 
@@ -301,37 +301,6 @@ class JobModule(BaseModule):
         if self.add_preview is None:
             return
         self.add_preview.preview_file(self.add_files[row] if 0 <= row < len(self.add_files) else None)
-
-    def _check_po_rfq_files(self, newly_added: List[str], files_store: List[str], list_widget) -> None:
-        """Scan newly-added files for PO/RFQ signals and prompt the user for each flagged file.
-
-        Only files whose extension matches the configured blueprint extensions are checked —
-        non-drawing files (.msg, .docx, etc.) are always documents and need no prompt.
-        """
-        bp_exts = [e.lower() for e in self.app_context.get_setting('blueprint_extensions', ['.pdf', '.dwg', '.dxf'])]
-        for path in list(newly_added):
-            if Path(path).suffix.lower() not in bp_exts:
-                continue
-            is_flagged, reason = classify_document(path)
-            if not is_flagged:
-                continue
-            filename = os.path.basename(path)
-            msg = QMessageBox(self._widget)
-            msg.setWindowTitle("PO / RFQ Detected")
-            msg.setIcon(QMessageBox.Icon.Warning)
-            msg.setText(f"'{filename}' appears to be a Purchase Order or RFQ.")
-            msg.setInformativeText(
-                f"Detected: {reason}\n\n"
-                "Save it as a document in the job folder, or remove it from the list?"
-            )
-            save_btn = msg.addButton("Save as Document", QMessageBox.ButtonRole.AcceptRole)
-            remove_btn = msg.addButton("Remove", QMessageBox.ButtonRole.RejectRole)
-            msg.setDefaultButton(save_btn)
-            msg.exec()
-            if msg.clickedButton() is remove_btn and path in files_store:
-                idx = files_store.index(path)
-                files_store.pop(idx)
-                list_widget.takeItem(idx)
 
     def handle_job_files(self, files: List[str]):
         """Add files to the job files list (Create New tab)"""
